# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List service requests
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: mine
          in: query
          required: false
          schema:
            type: boolean
        - name: providerId
          in: query
          required: false
          schema:
            type: integer
        - name: lat
          in: query
          required: false
          description: Latitude of the reference point for nearby filtering.
          schema:
            type: number
        - name: lng
          in: query
          required: false
          description: Longitude of the reference point for nearby filtering.
          schema:
            type: number
        - name: radiusKm
          in: query
          required: false
          description: Only return requests within this many kilometers of lat/lng.
          schema:
            type: number
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a service request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a request by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update a request (status, assignment, schedule)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{id}/offers:
    get:
      operationId: listRequestOffers
      tags: [offers]
      summary: List offers for a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
  /requests/{id}/contact:
    get:
      operationId: getRequestContact
      tags: [requests]
      summary: Get the other party's contact details (revealed only after the request is accepted)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact details of the counterparty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestContact"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Not available until the request is accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    RequestContact:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, provider]
          description: The role of the counterparty whose contact is returned.
...
    ServiceRequest:
      type: object
      required:
        [
          id,
          requestNumber,
          userId,
          skillId,
          status,
          includesSpareParts,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
        requestNumber:
          type: string
        userId:
          type: integer
        providerId:
          type: ["integer", "null"]
        skillId:
          type: integer
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        includesSpareParts:
          type: boolean
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        address:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
          enum: [on_site, null]
        status:
          type: string
          enum: [pending, active, in_progress, completed, cancelled]
        priceMin:
          type: ["integer", "null"]
        priceMax:
          type: ["integer", "null"]
        preferredTime:
          type: ["string", "null"]
          format: date-time
        scheduledTime:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RequestInput:
      type: object
      required: [skillId]
      properties:
        skillId:
          type: integer
        description:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        includesSpareParts:
          type: boolean
        lat:
          type: number
        lng:
          type: number
        address:
          type: string
        paymentMethod:
          type: string
          enum: [on_site]
        priceMin:
          type: integer
        priceMax:
          type: integer
        preferredTime:
          type: string
          format: date-time
        scheduledTime:
          type: string
          format: date-time
    RequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, in_progress, completed, cancelled]
        providerId:
          type: integer
        scheduledTime:
          type: string
          format: date-time
        priceMin:
          type: integer
        priceMax:
          type: integer
    Offer:
      type: object
      required: [id, requestId, providerId, price, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        providerId:
          type: integer
        price:
          type: integer
        message:
          type: ["string", "null"]
        availableTime:
          type: ["string", "null"]
          format: date-time
        estimatedDuration:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...